% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphavantagerxtra.R
\name{av_cache_retrieve}
\alias{av_cache_retrieve}
\title{av_cache_retrieve}
\usage{
av_cache_retrieve(insymbol)
}
\arguments{
\item{insymbol}{A ticker symbol e.g. AAPL}
}
\value{
tibble
}
\description{
Retrieve all cached reports for company
}
\examples{
av_cache_retrieve("AAPL")

}
