% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphavantagerxtra.R
\name{av_cache_retrieve_report}
\alias{av_cache_retrieve_report}
\title{av_cache_retrieve_report}
\usage{
av_cache_retrieve_report(data_set)
}
\arguments{
\item{data_set}{balance_sheet|cash_flow|overview|income_statement|earnings|ipo_calendar}
}
\value{
tibble
}
\description{
Retrieve cache file by report type.
}
\examples{
av_cache_retrieve_report("ipo_calendar")


}
