% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphavantagerxtra.R
\name{av_data_set_files}
\alias{av_data_set_files}
\title{av_data_set_files}
\usage{
av_data_set_files()
}
\value{
tibble
}
\description{
Cache file names. Primarily for internal use only.
}
\examples{
av_data_set_files()

}
