% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphavantagerxtra.R
\name{av_format}
\alias{av_format}
\title{av_format}
\usage{
av_format(data)
}
\value{
tibble
}
\description{
Primarily for internal use. Performs some data hygiene on
Alpha vantage API response and converts column names to snake case.
}
