% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphavantagerxtra.R
\name{av_get_annual_reports}
\alias{av_get_annual_reports}
\title{av_get_annual_reports}
\usage{
av_get_annual_reports(symbol)
}
\arguments{
\item{symbol}{A ticker symbol e.g. AAPL}
}
\value{
structure with 3 tibbles
}
\description{
Get primary financial documents for a given symbol including
Balance Sheet, Cash Flow,
}
\examples{
av_get_annual_reports("IBM")

}
