% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphavantagerxtra.R
\name{av_get_balance_sheet}
\alias{av_get_balance_sheet}
\title{av_get_balance_sheet}
\usage{
av_get_balance_sheet(symbol)
}
\arguments{
\item{symbol}{A ticker symbol e.g. FB}
}
\value{
tibble
}
\description{
Get Balance Sheet
}
\examples{
av_get_balance_sheet("FB")

}
