% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphavantagerxtra.R
\name{av_get_cash_flow}
\alias{av_get_cash_flow}
\title{av_get_cash_flow}
\usage{
av_get_cash_flow(symbol)
}
\arguments{
\item{symbol}{A ticker symbol e.g. JPM}
}
\value{
tibble
}
\description{
Get Cash Flow financial document
}
\examples{
av_get_cash_flow("JPM")

}
