% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphavantagerxtra.R
\name{av_get_earnings}
\alias{av_get_earnings}
\title{av_get_earnings}
\usage{
av_get_earnings(symbol)
}
\arguments{
\item{symbol}{A ticker symbol e.g. TWTR}
}
\value{
tibble
}
\description{
Get earnings
}
\examples{
av_get_earnings("TWTR")

}
