% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphavantagerxtra.R
\name{av_get_income_statement}
\alias{av_get_income_statement}
\title{av_get_income_statement}
\usage{
av_get_income_statement(symbol)
}
\arguments{
\item{symbol}{A ticker symbol e.g. MS}
}
\value{
tibble
}
\description{
Get Income statement for company
}
\examples{
av_get_income_statement("MS")

}
