% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphavantagerxtra.R
\name{av_get_overview}
\alias{av_get_overview}
\title{av_get_overview}
\usage{
av_get_overview(symbol)
}
\arguments{
\item{symbol}{A ticker symbol e.g. AMZN}
}
\value{
tibble
}
\description{
Get business overview
}
\examples{
av_get_overview("AMZN")

}
